Add RBAC authorization framework
Users are members of roles.  Objects (e.g. datasets or narratives) have
tags.  Policies define rules which allow a given role to take a set of
actions on any object with a given tag.

Currently the policies, tags, and roles are statically defined, but
these can easily become customized or user-defined in the future.

Cognito group memberships change from being equivalent to Nextstrain
Groups memberships to instead being authorization role groups, including
membership roles for Nextstrain Groups (viewers, editors, owners).

Completely replaces previous interim access control via visibleToUser()
methods.

The Charon API endpoints now return 401/403 instead of 404 to
unauthorized requests since their access control checks now go through
the app-wide error handler.  This centralization is an overall
improvement and doesn't leak existence of private paths any more than we
already were with other endpoints.

+++ src/endpoints/sources.js
@@ -12,7 +12,7 @@ const {promisify} = require("util");
 const zlib = require("zlib");
 const readStream = require("raw-body");
 
-const {AuthzDenied} = require("../exceptions");
+const authz = require("../authz");
 const {contentTypesProvided, contentTypesConsumed} = require("../negotiate");
 const {fetch, Request} = require("../fetch");
 const sources = require("../sources");
@@ -45,9 +45,7 @@ const setSource = (sourceName, argsExtractor = (req) => []) => (req, res, next) 
 
   res.vary("Accept");
 
-  if (!source.visibleToUser(req.user)) {
-    throw new AuthzDenied();
-  }
+  authz.assertAuthorized(req.user, authz.actions.Read, source);
 
   req.context.source = source;
   return next();
@@ -123,6 +121,8 @@ const canonicalizeDataset = (pathBuilder) => (req, res, next) => {
  * @type {expressMiddlewareAsync}
  */
 const ifDatasetExists = async (req, res, next) => {
+  authz.assertAuthorized(req.user, authz.actions.Read, req.context.dataset);
+
   if (!(await req.context.dataset.exists())) throw new NotFound();
   return next();
 };
@@ -209,6 +209,8 @@ const putDataset = contentTypesConsumed([
 const deleteResource = resourceExtractor => async (req, res) => {
   const resource = resourceExtractor(req);
 
+  authz.assertAuthorized(req.user, authz.actions.Write, resource);
+
   const method = "DELETE";
   const upstreamUrls = await Promise.all(resource.subresources().map(s => s.url(method)));
   const upstreamResponses = await Promise.all(upstreamUrls.map(url => fetch(url, {method})));
@@ -250,6 +252,8 @@ const setNarrative = (pathExtractor) => (req, res, next) => {
  * @type {expressMiddlewareAsync}
  */
 const ifNarrativeExists = async (req, res, next) => {
+  authz.assertAuthorized(req.user, authz.actions.Read, req.context.narrative);
+
   if (!(await req.context.narrative.exists())) throw new NotFound();
   return next();
 };
@@ -322,6 +326,9 @@ function pathParts(path = "") {
 function sendSubresource(subresourceExtractor) {
   return async (req, res) => {
     const subresource = subresourceExtractor(req);
+
+    authz.assertAuthorized(req.user, authz.actions.Read, subresource.resource);
+
     const subresourceUrl = await subresource.url();
 
     /* Proxy the data through us:
@@ -370,6 +377,8 @@ function receiveSubresource(subresourceExtractor) {
   return async (req, res) => {
     const method = "PUT";
     const subresource = subresourceExtractor(req);
+
+    authz.assertAuthorized(req.user, authz.actions.Write, subresource.resource);
 
     /* Proxy the data through us:
      *
